requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==24.2.0
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
deepspeed==0.14.4
diffusers==0.30.3
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
grpcio==1.67.1
hjson==3.1.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
ml_collections==1.0.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.1
ninja==1.11.1.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.19.2
open_clip_torch==2.29.0
opencv-python==4.10.0.84
optimum-quanto==0.2.2
packaging==24.1
pandas==2.2.3
peft==0.14.0
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.28.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
timm==1.0.11
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.5
transformers==4.43.3
triton==2.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.3
wcwidth==0.2.13
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.16.0
zipp==3.20.2
